backend/requirements/requirements-bionic.txt
Parsley==1.2
Twisted==17.9.0
acme==0.25.1
chardet==3.0.4
cffi==1.11.5
cryptography==2.1.4
idna==2.6
pgpy==0.4.3
pyOpenSSL==17.5.0
pyasn1==0.4.2
pyasn1_modules==0.2.1
pycparser==2.18
python_debian==0.1.32
python_gnupg==0.4.1
requests==2.18.4
scrypt==0.8.0
service_identity==16.0.0
six==1.11.0
sqlalchemy==1.1.11
transaction==1.4.3
txtorcon==0.19.3
zope.component==4.3.0
zope.event==4.2.0
zope.interface==4.3.2
